requirements.txt
python-telegram-bot==13.15
pybit==5.4.3
schedule